pyproject.toml
[tool.poetry]
name = "kyrcovay-three"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
psycopg2 = "^2.9.11"
python-dotenv = "^1.2.1"
requests = "^2.32.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
